bot/requirements.txt
python-telegram-bot>=20.0,<22.0
APScheduler>=3.10.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
